TARO/requirements.txt
cx_oracle
Django==4.2.11
djangorestframework
djangorestframework-xml
django-classy-tags>=0.9
django-cors-headers
django-filer>=1.3
django-filter
django-haystack
django-mptt>0.9
django_otp
django-sekizai>=1.0
django-sslserver
django-treebeard>=4.0,<5.0
drf-haystack
easy_thumbnails
html5lib>=1.0.1
lxml
markdown
paramiko==2.7.1
Pillow>=3.0
pysolr
qrcode
requests==2.24.0
tox
xmltodict
wait-for-it
whitenoise>=5.3.0
mysqlclient~=2.0.1
stopit